python/algorithms/cloud/awsS3_to_AzureStorage_migration/requirements.txt
boto3
azure-storage-blob
python-dotenv